# blob-store/api/spec/swagger.yml
openapi: 3.0.0
info:
    title: Blob Storage API
    description: API for storing and retrieving blobs of data
    version: 0.1.0
servers:
    - url: http://localhost:8081
      description: Local instance
paths:
  /api/v1/blobs/upload:
    post:
      summary: Upload blob
      operationId: uploadBlob
      parameters:
        - $ref: '#/components/parameters/xTraceID'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - format
              properties:
                file:
                  type: string
                  format: binary
                format:
                  type: string
      responses:
        201:
          $ref: "#/components/responses/UploadBlobSuccessResponse"
        400:
          $ref: "#/components/responses/ErrorBadRequest"
        500:
          $ref: "#/components/responses/ErrorInternal"


components:
  schemas:
    Error:
      type: object
      description: Object with error type and description
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
      example:
        type: "blob_not_provided"
        message: "Blob was not provided in request body"
  parameters:
    xTraceID:
      in: header
      name: X-Trace-ID
      description: Globally unique identifier of the request in UUID v4 format
      required: true
      schema:
        type: string
        format: uuid
  responses:
    UploadBlobSuccessResponse:
      description: Response to upload blob request
      content:
        application/json:
          schema:
            type: object
            properties:
              blob_id:
                type: string
            required:
              - blob_id
    ErrorBadRequest:
      description: Unable to process the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrorInternal:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"